refactor: separate linear and tree execiton entirely
